[WIP] Updated Contract, moved helper functions into contract
