+++ src/main/java/at/bitfire/jtx/JtxContract.kt
@@ -11,6 +11,15 @@ package at.bitfire.jtx
 import android.accounts.Account
 import android.net.Uri
 import android.provider.BaseColumns
+import android.util.Log
+import net.fortuna.ical4j.model.ParameterList
+import net.fortuna.ical4j.model.Property
+import net.fortuna.ical4j.model.PropertyList
+import net.fortuna.ical4j.model.parameter.XParameter
+import net.fortuna.ical4j.model.property.XProperty
+import org.json.JSONObject
+import java.lang.NullPointerException
+
 
 @Suppress("unused")
 object JtxContract {
@@ -34,9 +43,12 @@ object JtxContract {
     const val AUTHORITY = "at.techbee.jtx.provider"
 
     /** The version of this SyncContentProviderContract */
-    const val VERSION = 1
+    const val CONTRACT_VERSION = 1
 
-
+    /** Constructs an Uri for the Jtx Sync Adapter with the given Account
+     * @param [account] The account that should be appended to the Base Uri
+     * @return [Uri] with the appended Account
+     */
     fun Uri.asSyncAdapter(account: Account): Uri =
         buildUpon()
             .appendQueryParameter(CALLER_IS_SYNCADAPTER, "true")
@@ -44,9 +56,131 @@ object JtxContract {
             .appendQueryParameter(ACCOUNT_TYPE, account.type)
             .build()
 
+    /**
+     * This function takes a string and tries to parse it to a list of XParameters.
+     * This is the counterpart of getJsonStringFromXParameters(...)
+     * @param [string] that should be parsed
+     * @return The list of XParameter parsed from the string
+     */
+    fun getXParametersFromJson(string: String): List<XParameter> {
+
+        val jsonObject = JSONObject(string)
+        val xparamList = mutableListOf<XParameter>()
+        for (i in 0 until jsonObject.length()) {
+            val names = jsonObject.names() ?: break
+            val xparamName = names[i]?.toString() ?: break
+            val xparamValue = jsonObject.getString(xparamName).toString()
+            if(xparamName.isNotBlank() && xparamValue.isNotBlank()) {
+                val xparam = XParameter(xparamName, xparamValue)
+                xparamList.add(xparam)
+            }
+        }
+        return xparamList
+    }
+
+    /**
+     * This function takes a string and tries to parse it to a list of XProperty.
+     * This is the counterpart of getJsonStringFromXProperties(...)
+     * @param [string] that should be parsed
+     * @return The list of XProperty parsed from the string
+     */
+    fun getXPropertyListFromJson(string: String): PropertyList<Property> {
+
+        val propertyList = PropertyList<Property>()
+
+        if(string.isBlank())
+            return propertyList
+
+        try {
+            val jsonObject = JSONObject(string)
+            for (i in 0 until jsonObject.length()) {
+                val names = jsonObject.names() ?: break
+                val propertyName = names[i]?.toString() ?: break
+                val propertyValue = jsonObject.getString(propertyName).toString()
+                if (propertyName.isNotBlank() && propertyValue.isNotBlank()) {
+                    val prop = XProperty(propertyName, propertyValue)
+                    propertyList.add(prop)
+                }
+            }
+        } catch (e: NullPointerException) {
+            Log.w("XPropertyList", "Error parsing x-property-list $string\n$e")
+        }
+        return propertyList
+    }
+
+
+    /**
+     * Takes a Parameter List and returns a Json String to be saved in a DB field.
+     * This is the counterpart to getXParameterFromJson(...)
+     * @param [parameters] The ParameterList that should be transformed into a Json String
+     * @return The generated Json object as a [String]
+     */
+    fun getJsonStringFromXParameters(parameters: ParameterList?): String? {
+
+        if(parameters == null)
+            return null
+
+        val jsonObject = JSONObject()
+        parameters.forEach { parameter ->
+            jsonObject.put(parameter.name, parameter.value)
+        }
+        return if(jsonObject.length() == 0)
+            null
+        else
+            jsonObject.toString()
+    }
+
+    /**
+     * Takes a Property List and returns a Json String to be saved in a DB field.
+     * This is the counterpart to getXPropertyListFromJson(...)
+     * @param [propertyList] The PropertyList that should be transformed into a Json String
+     * @return The generated Json object as a [String]
+     */
+    fun getJsonStringFromXProperties(propertyList: PropertyList<*>?): String? {
+
+        if(propertyList == null)
+            return null
+
+        val jsonObject = JSONObject()
+        propertyList.forEach { property ->
+            jsonObject.put(property.name, property.value)
+        }
+        return if(jsonObject.length() == 0)
+            null
+        else
+            jsonObject.toString()
+    }
+
+
+    /**
+     * Some date fields in jtx Board are saved as a list of Long values separated by commas.
+     * This applies for example to the exdate for recurring events.
+     * This function takes a string and tries to parse it to a list of long values (timestamps)
+     * @param [string] that should be parsed
+     * @return a [MutableList<Long>] with the timestamps parsed from the string
+     *
+     */
+    fun getLongListFromString(string: String): MutableList<Long> {
+
+        val stringList = string.split(",")
+        val longList = mutableListOf<Long>()
+
+        stringList.forEach {
+            try {
+                longList.add(it.toLong())
+            } catch (e: NumberFormatException) {
+                Log.w("getLongListFromString", "String could not be cast to Long ($it)")
+                return@forEach
+            }
+        }
+        return longList
+    }
+
 
     @Suppress("unused")
     object JtxICalObject {
+
+
 
         /** The name of the the content URI for IcalObjects.
          * This is a general purpose table containing general columns
@@ -56,26 +190,28 @@ object JtxContract {
         /** The content uri of the ICalObject table */
         val CONTENT_URI: Uri by lazy { Uri.parse("content://$AUTHORITY/$CONTENT_URI_PATH") }
 
-
         /** Constant to define all day values (for dtstart, due, completed timezone fields */
         const val TZ_ALLDAY = "ALLDAY"
 
-
         /** The name of the ID column.
-         * This is the unique identifier of an ICalObject
+         * This is the unique identifier of an ICalObject.
          * Type: [Long]*/
         const val ID = BaseColumns._ID
 
         /** The column for the module.
-         * This is an internal differentiation for JOURNAL, NOTE and TODO
-         * provided in the enum [Module]
+         * This is an internal differentiation for JOURNAL, NOTE and TODO as provided in the enum [Module].
+         * The Module does not need to be set. On import it will be derived from the component from the [Component] (Todo or Journal/Note) and if a
+         * dtstart is present or not (Journal or Note). If the module was set, it might be overwritten on import. In this sense also
+         * unknown values are overwritten.
+         * Use e.g. Module.JOURNAL.name for a correct String value in this field.
          * Type: [String]
          */
         const val MODULE = "module"
 
-        /* The names of all the other columns  */
-        /** The column for the component based on the values
-         * provided in the enum [Component]
+        /***** The names of all the other columns  *****/
+        /** The column for the component based on the values provided in the enum [Component].
+         * Use e.g. Component.VTODO.name for a correct String value in this field.
+         * If no Component is provided on Insert, an IllegalArgumentException is thrown
          * Type: [String]
          */
         const val COMPONENT = "component"
@@ -96,6 +232,7 @@ object JtxContract {
 
         /**
          * Purpose:  This column/property specifies when the calendar component begins.
+         * This value is stored as UNIX timestamp (milliseconds).
          * The corresponding timezone is stored in [DTSTART_TIMEZONE].
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.2.4]
          * Type: [Long]
@@ -105,6 +242,11 @@ object JtxContract {
         /**
          * Purpose:  This column/property specifies the timezone of when the calendar component begins.
          * The corresponding datetime is stored in [DTSTART].
+         * The value of a timezone can be:
+         * 1. the id of a Java timezone to represent the given timezone.
+         * 2. null to represent floating time.
+         * 3. the value of [TZ_ALLDAY] to represent an all-day entry.
+         * If an invalid value is passed, the Timezone is ignored and interpreted as UTC.
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.2.4]
          * Type: [String]
          */
@@ -112,6 +254,7 @@ object JtxContract {
 
         /**
          * Purpose:  This column/property specifies when the calendar component ends.
+         * This value is stored as UNIX timestamp (milliseconds).
          * The corresponding timezone is stored in [DTEND_TIMEZONE].
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.2.4]
          * Type: [Long]
@@ -121,6 +264,7 @@ object JtxContract {
         /**
          * Purpose:  This column/property specifies the timezone of when the calendar component ends.
          * The corresponding datetime is stored in [DTEND].
+         * See [DTSTART_TIMEZONE] for information about the timezone handling
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.2.2]
          * Type: [String]
          */
@@ -129,6 +273,8 @@ object JtxContract {
         /**
          * Purpose:  This property defines the overall status or confirmation for the calendar component.
          * The possible values of a status are defined in [StatusTodo] for To-Dos and in [StatusJournal] for Notes and Journals
+         * If no valid Status is used, the value is taken and will be shown as-is.
+         * Also null-value is allowed.
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.1.11]
          * Type: [String]
          */
@@ -137,6 +283,9 @@ object JtxContract {
         /**
          * Purpose:  This property defines the access classification for a calendar component.
          * The possible values of a status are defined in the enum [Classification].
+         * Use e.g. Classification.PUBLIC.name to put a correct String value in this field.
+         * If no valid Classification is used, the value is taken and will be shown as-is.
+         * Also null-value is allowed.
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.1.11]
          * Type: [String]
          */
@@ -162,7 +311,7 @@ object JtxContract {
          * This property is split in the fields [GEO_LAT] for the latitude
          * and [GEO_LONG] for the longitude coordinates.
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.1.6]
-         * Type: [Double]
+         * Type: [Float]
          */
         const val GEO_LAT = "geolat"
 
@@ -175,11 +324,10 @@ object JtxContract {
          */
         const val GEO_LONG = "geolong"
 
-
         /**
          * Purpose:  This property defines the intended venue for the activity defined by a calendar component.
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.1.7]
-         * Type: [String]
+         * Type: [Double]
          */
         const val LOCATION = "location"
 
@@ -192,6 +340,8 @@ object JtxContract {
 
         /**
          * Purpose:  This property is used by an assignee or delegatee of a to-do to convey the percent completion of a to-do to the "Organizer".
+         * The value must either be null or between 0 and 100. The value 0 is interpreted as null.
+         * If a value outside of the boundaries (0-100) is passed, the value is reset to null.
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.1.8]
          * Type: [Int]
          */
@@ -199,13 +349,15 @@ object JtxContract {
 
         /**
          * Purpose:  This property defines the relative priority for a calendar component.
-         * See [https://tools.ietf.org/html/rfc5545#section-3.8.1.9]
+         * See [https://tools.ietf.org/html/rfc5545#section-3.8.1.9]. The priority can be null or between 0 and 9.
+         * Values outside of these boundaries are accepted but internally not mapped to a string resource.
          * Type: [Int]
          */
         const val PRIORITY = "priority"
 
         /**
          * Purpose:  This property defines the date and time that a to-do is expected to be completed.
+         * This value is stored as UNIX timestamp (milliseconds).
          * The corresponding timezone is stored in [DUE_TIMEZONE].
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.2.3]
          * Type: [Long]
@@ -215,6 +367,7 @@ object JtxContract {
         /**
          * Purpose:  This column/property specifies the timezone of when a to-do is expected to be completed.
          * The corresponding datetime is stored in [DUE].
+         * See [DTSTART_TIMEZONE] for information about the timezone handling
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.2.2]
          * Type: [String]
          */
@@ -222,6 +375,7 @@ object JtxContract {
 
         /**
          * Purpose:  This property defines the date and time that a to-do was actually completed.
+         * This value is stored as UNIX timestamp (milliseconds).
          * The corresponding timezone is stored in [COMPLETED_TIMEZONE].
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.2.1]
          * Type: [Long]
@@ -230,7 +384,8 @@ object JtxContract {
 
         /**
          * Purpose:  This column/property specifies the timezone of when a to-do was actually completed.
-         * The corresponding datetime is stored in [DUE].
+         * The corresponding datetime is stored in [COMPLETED].
+         * See [DTSTART_TIMEZONE] for information about the timezone handling
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.2.1]
          * Type: [String]
          */
@@ -238,6 +393,11 @@ object JtxContract {
 
         /**
          * Purpose:  This property specifies a positive duration of time.
+         * See [DTSTART_TIMEZONE] for information about the timezone handling
+         * The string representation follows the notation as given in RFC-5545
+         * for the value type duration: [https://datatracker.ietf.org/doc/html/rfc5545#section-3.3.6]
+         * Exampe: "P15DT5H0M20S". This field is currently not in use. If present, the user would
+         * see a notification that a duration cannot be processed and will be overwritten if the entry is changed.
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.2.5]
          * Type: [String]
          */
@@ -247,6 +407,10 @@ object JtxContract {
         /**
          * Purpose:  This property defines a rule or repeating pattern for recurring events,
          * to-dos, journal entries, or time zone definitions.
+         * The representation of the RRULE follows the value type RECUR of RFC-5545 as given in
+         * [https://datatracker.ietf.org/doc/html/rfc5545#section-3.3.10]
+         * For example: "FREQ=DAILY;COUNT=10"
+         * See also [https://datatracker.ietf.org/doc/html/rfc5545#section-3.8.5.3].
          * Type: [String]
          */
         const val RRULE = "rrule"
@@ -254,14 +418,18 @@ object JtxContract {
         /**
          * Purpose:  This property defines the list of DATE-TIME values for
          * recurring events, to-dos, journal entries, or time zone definitions.
-         * Type: [String], contains a list of comma-separated date values as Long
+         * Type: [String], contains a list of comma-separated date values as
+         * UNIX timestamps (milliseconds) e.g. "1636751475000,1636837875000".
+         * Invalid values that cannot be transformed to [Long] are ignored.
          */
         const val RDATE = "rdate"
 
         /**
          * Purpose:  This property defines the list of DATE-TIME exceptions for
          * recurring events, to-dos, journal entries, or time zone definitions.
-         * Type: [String], contains a list of comma-separated date values as Long
+         * Type: [String], contains a list of comma-separated date values as
+         * UNIX timestamps (milliseconds) e.g. "1636751475000,1636837875000".
+         * Invalid values that cannot be transformed to [Long] are ignored.
          */
         const val EXDATE = "exdate"
 
@@ -270,17 +438,18 @@ object JtxContract {
          * "SEQUENCE" properties to identify a specific instance of a
          * recurring "VEVENT", "VTODO", or "VJOURNAL" calendar component.
          * The property value is the original value of the "DTSTART" property
-         * of the recurrence instance.
+         * of the recurrence instance, ie. a DATE or DATETIME value e.g. "20211101T160000".
+         * Must be null for non-recurring and original events from which recurring events are derived.
+         * Type: [String]
          */
         const val RECURID = "recurid"
 
         /**
          * Stores the reference to the original event from which the recurring event was derived.
-         * This value is NULL for the orignal event.
-         * Type: [Long]
+         * This value is NULL for the orignal event or if the event is not recurring
+         * Type: [Long], References [JtxICalObject.ID]
          */
         const val RECUR_ORIGINALICALOBJECTID = "original_id"
-
 
         /**
          * Purpose:  This property defines the persistent, globally unique identifier for the calendar component.
@@ -291,7 +460,7 @@ object JtxContract {
 
         /**
          * Purpose:  This property specifies the date and time that the calendar information
-         * was created by the calendar user agent in the calendar store.
+         * was created by the calendar user agent in the calendar store and is stored as UNIX timestamp (milliseconds).
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.7.1]
          * Type: [Long]
          */
@@ -304,7 +473,7 @@ object JtxContract {
          * an iCalendar object that doesn't specify a "METHOD" property, this
          * property specifies the date and time that the information
          * associated with the calendar component was last revised in the
-         * calendar store.
+         * calendar store. It is saved as UNIX timestamp (milliseconds).
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.7.2]
          * Type: [Long]
          */
@@ -313,6 +482,7 @@ object JtxContract {
         /**
          * Purpose:  This property specifies the date and time that the information associated
          * with the calendar component was last revised in the calendar store.
+         * It is saved as UNIX timestamp (milliseconds).
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.7.3]
          * Type: [Long]
          */
@@ -320,6 +490,7 @@ object JtxContract {
 
         /**
          * Purpose:  This property defines the revision sequence number of the calendar component within a sequence of revisions.
+         * If no sequence is present, it is automatically initialized with 0
          * See [https://tools.ietf.org/html/rfc5545#section-3.8.7.4]
          * Type: [Int]
          */
@@ -327,20 +498,20 @@ object JtxContract {
 
         /**
          * Purpose:  This property specifies a color used for displaying the calendar, event, todo, or journal data.
-         * See [https://tools.ietf.org/html/rfc7986#section-5.9]
-         * The expected String is a String that can be parsed by Color.parseColor(...)
-         * Type: [String]
+         * See [https://tools.ietf.org/html/rfc7986#section-5.9].
+         * The color is represented as Int-value as described in [https://developer.android.com/reference/android/graphics/Color#color-ints]
+         * Type: [Int]
          */
         const val COLOR = "color"
 
         /**
          * Purpose:  This column is the foreign key to the [JtxCollection].
-         * Type: [Long]
+         * Type: [Long], references [JtxCollection.ID]
          */
         const val ICALOBJECT_COLLECTIONID = "collectionId"
 
         /**
-         * Purpose:  This column defines if a local collection was changed that is supposed to be synchronised.
+         * Purpose:  This column defines if the local ICalObject was changed and that it is supposed to be synchronised.
          * Type: [Boolean]
          */
         const val DIRTY = "dirty"
@@ -352,7 +523,8 @@ object JtxContract {
         const val DELETED = "deleted"
 
         /**
-         * Purpose:  filename of the synched entry (*.ics), only relevant for synched entries through sync-adapter
+         * Purpose:  The remote file name of the synchronized entry (*.ics), only relevant for synchronized
+         * entries through the sync-adapter
          * Type: [String]
          */
         const val FILENAME = "filename"
@@ -375,33 +547,59 @@ object JtxContract {
          */
         const val FLAGS = "flags"
 
+        /**
+         * Purpose:  To specify other properties for the ICalObject.
+         * This is especially used for additional properties that cannot be put into another field, especially relevant for the synchronization
+         * The Properties are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXProperties(PropertyList<*> that returns a Json String from the property list
+         * to be stored in this other field. The counterpart to this function is
+         * getXPropertyListFromJson(String) that returns a PropertyList from a Json that was created with getJsonStringFromXProperties()
+         * Type: [String]
+         */
+        const val OTHER = "other"
 
-        /** This enum class defines the possible values for the attribute status of an [JtxICalObject] for Journals/Notes */
+
+
+
+        /**
+         * This enum class defines the possible values for the attribute status of an [JtxICalObject] for Journals/Notes
+         *  Use its name when the string representation is required, e.g. StatusJournal.DRAFT.name.
+         */
         enum class StatusJournal {
             DRAFT, FINAL, CANCELLED
         }
 
-        /** This enum class defines the possible values for the attribute status of an [JtxICalObject] for Todos */
+        /**
+         * This enum class defines the possible values for the attribute status of an [JtxICalObject] for Todos
+         * Use its name when the string representation is required, e.g. StatusTodo.`NEEDS-ACTION`.name.
+         */
         enum class StatusTodo {
             `NEEDS-ACTION`, COMPLETED, `IN-PROCESS`, CANCELLED
         }
 
-        /** This enum class defines the possible values for the attribute classification of an [JtxICalObject]  */
+        /**
+         * This enum class defines the possible values for the attribute classification of an [JtxICalObject]
+         * Use its name when the string representation is required, e.g. Classification.PUBLIC.name.
+         */
         enum class Classification {
             PUBLIC, PRIVATE, CONFIDENTIAL
         }
 
-        /** This enum class defines the possible values for the attribute component of an [JtxICalObject]  */
+        /**
+         * This enum class defines the possible values for the attribute component of an [JtxICalObject]
+         * Use its name when the string representation is required, e.g. Component.VJOURNAL.name.
+         */
         enum class Component {
             VJOURNAL, VTODO
         }
 
-        /** This enum class defines the possible values for the attribute module of an [JtxICalObject]  */
+        /**
+         * This enum class defines the possible values for the attribute module of an [JtxICalObject]
+         * Use its name when the string representation is required, e.g. Module.JOURNAL.name.
+         */
         enum class Module {
             JOURNAL, NOTE, TODO
         }
-
-
     }
 
 
@@ -422,21 +620,27 @@ object JtxContract {
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
-         * Type: [Long] */
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ICALOBJECT_ID = "icalObjectId"
 
 
-        /* The names of all the other columns  */
-
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This value type is used to identify properties that contain a calendar user address (in this case of the attendee).
-         * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.3.3]
+         * This is usually an e-mail address as defined in [https://datatracker.ietf.org/doc/html/rfc5545#section-3.3.3].
+         * The value should be passed as String containing the URI with "mailto:", for example: "mailto:jane_doe@example.com"
+         * See also [https://tools.ietf.org/html/rfc5545#section-3.8.4.1].
          * Type: [String]
          */
         const val CALADDRESS = "caladdress"
 
         /**
          * Purpose:  To identify the type of calendar user specified by the property in this case for the attendee.
+         * The possible values are defined in the enum [Cutype].
+         * Use e.g. Cutype.INDIVIDUAL.name to put a correct String value in this field.
+         * If no value is passed for the Cutype, the Cutype will be interpreted as INDIVIDUAL as according to the RFC.
+         * Other values are accepted and treated as UNKNOWN.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.2.3]
          * Type: [String]
          */
@@ -444,7 +648,6 @@ object JtxContract {
 
         /**
          * Purpose:  To specify the group or list membership of the calendar user specified by the property in this case for the attendee.
-         * The possible values are defined in the enum [Cutype]
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.2.11]
          * Type: [String]
          */
@@ -453,6 +656,9 @@ object JtxContract {
         /**
          * Purpose:  To specify the participation role for the calendar user specified by the property in this case for the attendee.
          * The possible values are defined in the enum [Role]
+         * Use e.g. Role.CHAIR.name to put a correct String value in this field.
+         * If no value (null) is passed for the Role, it will be interpreted as REQ-PARTICIPANT as according to the RFC.
+         * Other values are accepted and treated as REQ-PARTICIPANTs.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.2.16]
          * Type: [String]
          */
@@ -484,28 +690,33 @@ object JtxContract {
         /**
          * Purpose:  To specify the calendar users that have delegated their participation to the calendar user specified by the property
          * in this case for the attendee.
-         * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.2.4]
+         * This is usually an e-mail address as defined in [https://datatracker.ietf.org/doc/html/rfc5545#section-3.3.3].
+         * The value should be passed as String containing the URI with "mailto:", for example: "mailto:jane_doe@example.com"
+         * See also [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.2.4]
          * Type: [String]
          */
         const val DELEGATEDFROM = "delegatedfrom"
 
         /**
          * Purpose:  To specify the calendar user that is acting on behalf of the calendar user specified by the property in this case for the attendee.
-         * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.2.18]
+         * This is usually an e-mail address as defined in [https://datatracker.ietf.org/doc/html/rfc5545#section-3.3.3].
+         * The value should be passed as String containing the URI with "mailto:", for example: "mailto:jane_doe@example.com"
+         * See also [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://datatracker.ietf.org/doc/html/rfc5545#section-3.2.18]
          * Type: [String]
          */
         const val SENTBY = "sentby"
 
         /**
          * Purpose:  To specify the common name to be associated with the calendar user specified by the property in this case for the attendee.
-         * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.2.18]
+         * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://datatracker.ietf.org/doc/html/rfc5545#section-3.2.2]
          * Type: [String]
          */
         const val CN = "cn"
 
         /**
          * Purpose:  To specify reference to a directory entry associated with the calendar user specified by the property in this case for the attendee.
-         * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.2.2]
+         * Expected is an URI as defined in [https://datatracker.ietf.org/doc/html/rfc5545#section-3.2.6].
+         * The value should be passed as String, e.g. "ldap://example.com:6666/o=ABC%20Industries,c=US???(cn=Jim%20Dolittle)"
          * Type: [String]
          */
         const val DIR = "dir"
@@ -513,29 +724,37 @@ object JtxContract {
         /**
          * Purpose:  To specify the language for text values in a property or property parameter, in this case of the attendee.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1] and [https://tools.ietf.org/html/rfc5545#section-3.2.10]
+         * Language-Tag as defined in RFC5646, e.g. "en:Germany"
          * Type: [String]
          */
         const val LANGUAGE = "language"
 
         /**
-         * Purpose:  To specify other properties for the attendee.
+         * Purpose:  To specify other parameters for the attendee.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.1]
+         * The Parameters are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXParameters(ParameterList?) that returns a Json String from the parameter list
+         * to be stored in this other field. The counterpart to this function is
+         * getXParameterFromJson(String) that returns a list of XParameters from a Json that was created with getJsonStringFromXParameters(...)
          * Type: [String]
          */
         const val OTHER = "other"
 
 
-        /** This enum class defines the possible values for the attribute Cutype of an [JtxAttendee]  */
+        /**
+         * This enum class defines the possible values for the attribute Cutype of an [JtxAttendee]
+         * Use its name when the string representation is required, e.g. Cutype.INDIVIDUAL.name.
+         */
         enum class Cutype {
             INDIVIDUAL, GROUP, RESOURCE, ROOM, UNKNOWN
         }
-
-        /** This enum class defines the possible values for the attribute Role of an [JtxAttendee]  */
+        /**
+         * This enum class defines the possible values for the attribute Role of an [JtxAttendee]
+         * Use its name when the string representation is required, e.g. Role.`REQ-PARTICIPANT`.name.
+         */
         enum class Role {
             CHAIR, `REQ-PARTICIPANT`, `OPT-PARTICIPANT`, `NON-PARTICIPANT`
         }
-
-
     }
 
     @Suppress("unused")
@@ -555,11 +774,12 @@ object JtxContract {
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
-         * Type: [Long] */
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ICALOBJECT_ID = "icalObjectId"
 
 
-/* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This property defines the name of the category for a calendar component.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.2]
@@ -570,6 +790,7 @@ object JtxContract {
         /**
          * Purpose:  To specify the language for text values in a property or property parameter, in this case of the category.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.2] and [https://tools.ietf.org/html/rfc5545#section-3.2.10]
+         * Language-Tag as defined in RFC5646, e.g. "en:Germany"
          * Type: [String]
          */
         const val LANGUAGE = "language"
@@ -577,6 +798,10 @@ object JtxContract {
         /**
          * Purpose:  To specify other properties for the category.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.2]
+         * The Parameters are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXParameters(ParameterList?) that returns a Json String from the parameter list
+         * to be stored in this other field. The counterpart to this function is
+         * getXParameterFromJson(String) that returns a list of XParameters from a Json that was created with getJsonStringFromXParameters(...)
          * Type: [String]
          */
         const val OTHER = "other"
@@ -599,11 +824,12 @@ object JtxContract {
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
-         * Type: [Long] */
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ICALOBJECT_ID = "icalObjectId"
 
 
-/* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This property specifies non-processing information intended to provide a comment to the calendar user.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.4]
@@ -614,6 +840,7 @@ object JtxContract {
         /**
          * Purpose:  To specify the language for text values in a property or property parameter, in this case of the comment.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.4]
+         * Language-Tag as defined in RFC5646, e.g. "en:Germany"
          * Type: [String]
          */
         const val ALTREP = "altrep"
@@ -628,6 +855,10 @@ object JtxContract {
         /**
          * Purpose:  To specify other properties for the comment.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.4]
+         * The Parameters are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXParameters(ParameterList?) that returns a Json String from the parameter list
+         * to be stored in this other field. The counterpart to this function is
+         * getXParameterFromJson(String) that returns a list of XParameters from a Json that was created with getJsonStringFromXParameters(...)
          * Type: [String]
          */
         const val OTHER = "other"
@@ -650,11 +881,12 @@ object JtxContract {
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
-         * Type: [Long] */
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ICALOBJECT_ID = "icalObjectId"
 
 
-/* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This property defines the name of the contact for a calendar component.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.2]
@@ -672,6 +904,7 @@ object JtxContract {
         /**
          * Purpose:  To specify the language for text values in a property or property parameter, in this case of the contact.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.2] and [https://tools.ietf.org/html/rfc5545#section-3.2.10]
+         * Language-Tag as defined in RFC5646, e.g. "en:Germany"
          * Type: [String]
          */
         const val LANGUAGE = "language"
@@ -679,6 +912,10 @@ object JtxContract {
         /**
          * Purpose:  To specify other properties for the contact.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.2]
+         * The Parameters are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXParameters(ParameterList?) that returns a Json String from the parameter list
+         * to be stored in this other field. The counterpart to this function is
+         * getXParameterFromJson(String) that returns a list of XParameters from a Json that was created with getJsonStringFromXParameters(...)
          * Type: [String]
          */
         const val OTHER = "other"
@@ -702,14 +939,17 @@ object JtxContract {
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
-         * Type: [Long] */
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ICALOBJECT_ID = "icalObjectId"
 
 
-/* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This value type is used to identify properties that contain a calendar user address (in this case of the organizer).
-         * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.3] and [https://tools.ietf.org/html/rfc5545#section-3.3.3]
+         * This is usually an e-mail address as defined in [https://datatracker.ietf.org/doc/html/rfc5545#section-3.3.3].
+         * The value should be passed as String containing the URI with "mailto:", for example: "mailto:jane_doe@example.com"
+         * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.3].
          * Type: [String]
          */
         const val CALADDRESS = "caladdress"
@@ -723,6 +963,8 @@ object JtxContract {
 
         /**
          * Purpose:  To specify reference to a directory entry associated with the calendar user specified by the property in this case for the organizer.
+         * Expected is an URI as defined in [https://datatracker.ietf.org/doc/html/rfc5545#section-3.2.6].
+         * The value should be passed as String, e.g. "ldap://example.com:6666/o=ABC%20Industries,c=US???(cn=Jim%20Dolittle)"
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.3] and [https://tools.ietf.org/html/rfc5545#section-3.2.2]
          * Type: [String]
          */
@@ -730,7 +972,9 @@ object JtxContract {
 
         /**
          * Purpose:  To specify the calendar user that is acting on behalf of the calendar user specified by the property in this case for the organizer.
-         * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.3] and [https://tools.ietf.org/html/rfc5545#section-3.2.18]
+         * This is usually an e-mail address as defined in [https://datatracker.ietf.org/doc/html/rfc5545#section-3.3.3].
+         * The value should be passed as String containing the URI with "mailto:", for example: "mailto:jane_doe@example.com"
+         * See also [https://tools.ietf.org/html/rfc5545#section-3.8.4.3] and [https://datatracker.ietf.org/doc/html/rfc5545#section-3.2.18]
          * Type: [String]
          */
         const val SENTBY = "sentbyparam"
@@ -738,6 +982,7 @@ object JtxContract {
         /**
          * Purpose:  To specify the language for text values in a property or property parameter, in this case of the organizer.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.3] and [https://tools.ietf.org/html/rfc5545#section-3.2.10]
+         * Language-Tag as defined in RFC5646, e.g. "en:Germany"
          * Type: [String]
          */
         const val LANGUAGE = "language"
@@ -745,6 +990,10 @@ object JtxContract {
         /**
          * Purpose:  To specify other properties for the organizer.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.3]
+         * The Parameters are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXParameters(ParameterList?) that returns a Json String from the parameter list
+         * to be stored in this other field. The counterpart to this function is
+         * getXParameterFromJson(String) that returns a list of XParameters from a Json that was created with getJsonStringFromXParameters(...)
          * Type: [String]
          */
         const val OTHER = "other"
@@ -770,7 +1019,8 @@ object JtxContract {
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
-         * Type: [Long] */
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ICALOBJECT_ID = "icalObjectId"
 
         /** The name of the second Foreign Key Column of the related IcalObject
@@ -779,7 +1029,7 @@ object JtxContract {
         const val LINKEDICALOBJECT_ID = "linkedICalObjectId"
 
 
-        /* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This property is used to represent a relationship or reference between one calendar component and another.
          * The text gives the UID of the related calendar entry.
@@ -791,7 +1041,9 @@ object JtxContract {
         /**
          * Purpose:  To specify the type of hierarchical relationship associated
          * with the calendar component specified by the property.
-         * The possible relationship types are defined in the enum [Reltype]
+         * The possible relationship types are defined in the enum [Reltype].
+         * Use e.g. Reltype.PARENT.name to put a correct String value in this field.
+         * Other values and null-values are allowed, but will not be processed by the app.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.5] and [https://tools.ietf.org/html/rfc5545#section-3.2.15]
          * Type: [String]
          */
@@ -800,12 +1052,19 @@ object JtxContract {
         /**
          * Purpose:  To specify other properties for the related-to.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.4.5]
+         * The Parameters are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXParameters(ParameterList?) that returns a Json String from the parameter list
+         * to be stored in this other field. The counterpart to this function is
+         * getXParameterFromJson(String) that returns a list of XParameters from a Json that was created with getJsonStringFromXParameters(...)
          * Type: [String]
          */
         const val OTHER = "other"
 
 
-        /** This enum class defines the possible values for the attribute Reltype of an [JtxRelatedto]  */
+        /**
+         * This enum class defines the possible values for the attribute Reltype of an [JtxRelatedto].
+         * Use its name when the string representation is required, e.g. Reltype.PARENT.name.
+         */
         enum class Reltype {
             PARENT, CHILD, SIBLING
         }
@@ -828,11 +1087,12 @@ object JtxContract {
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
-         * Type: [Long] */
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ICALOBJECT_ID = "icalObjectId"
 
 
-/* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This property defines the name of the resource for a calendar component.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.10]
@@ -843,6 +1103,7 @@ object JtxContract {
         /**
          * Purpose:  To specify an alternate text representation for the property value, in this case of the resource.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.4]
+         * Language-Tag as defined in RFC5646, e.g. "en:Germany"
          * Type: [String]
          */
         const val LANGUAGE = "language"
@@ -850,6 +1111,10 @@ object JtxContract {
         /**
          * Purpose:  To specify other properties for the resource.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.10]
+         * The Parameters are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXParameters(ParameterList?) that returns a Json String from the parameter list
+         * to be stored in this other field. The counterpart to this function is
+         * getXParameterFromJson(String) that returns a list of XParameters from a Json that was created with getJsonStringFromXParameters(...)
          * Type: [String]
          */
         const val OTHER = "other"
@@ -884,7 +1149,7 @@ object JtxContract {
          * Type: [Long]*/
         const val ID = BaseColumns._ID
 
-        /* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This column/property defines the url of the collection.
          * Type: [String]
@@ -912,6 +1177,7 @@ object JtxContract {
         /**
          * Purpose:  This column/property defines the color of the collection items.
          * This color can also be overwritten by the color in an ICalObject.
+         * The color is represented as Int-value as described in [https://developer.android.com/reference/android/graphics/Color#color-ints]
          * Type: [Int]
          */
         const val COLOR = "color"
@@ -977,11 +1243,12 @@ object JtxContract {
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
-         * Type: [Long] */
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ICALOBJECT_ID = "icalObjectId"
 
 
-/* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This property specifies the uri of an attachment.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.1]
@@ -1006,6 +1273,10 @@ object JtxContract {
         /**
          * Purpose:  To specify other properties for the attachment.
          * see [https://tools.ietf.org/html/rfc5545#section-3.8.1.1]
+         * The Parameters are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXParameters(ParameterList?) that returns a Json String from the parameter list
+         * to be stored in this other field. The counterpart to this function is
+         * getXParameterFromJson(String) that returns a list of XParameters from a Json that was created with getJsonStringFromXParameters(...)
          * Type: [String]
          */
         const val OTHER = "other"
@@ -1025,7 +1296,8 @@ object JtxContract {
 
         /** The name of the ID column for attachments.
          * This is the unique identifier of an Attachment
-         * Type: [Long]*/
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
@@ -1042,9 +1314,13 @@ object JtxContract {
         const val ALARM_ATTACH = "attach"
 
 
-        /* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This property stores the unknown value as json
+         * The Parameters are stored as JSON. There are two helper functions provided:
+         * getJsonStringFromXParameters(ParameterList?) that returns a Json String from the parameter list
+         * to be stored in this other field. The counterpart to this function is
+         * getXParameterFromJson(String) that returns a list of XParameters from a Json that was created with getJsonStringFromXParameters(...)
          * Type: [String]
          */
         const val ALARM_OTHER = "other"
@@ -1062,7 +1338,8 @@ object JtxContract {
 
         /** The name of the ID column for attachments.
          * This is the unique identifier of an Attachment
-         * Type: [Long]*/
+         * Type: [Long], references [JtxICalObject.ID]
+         */
         const val ID = BaseColumns._ID
 
         /** The name of the Foreign Key Column for IcalObjects.
@@ -1070,7 +1347,7 @@ object JtxContract {
         const val ICALOBJECT_ID = "icalObjectId"
 
 
-        /* The names of all the other columns  */
+        /***** The names of all the other columns  *****/
         /**
          * Purpose:  This property stores the unknown value as json
          * Type: [String]
@@ -1078,4 +1355,3 @@ object JtxContract {
         const val UNKNOWN_VALUE = "value"
     }
 }
-
